+++ betty/demo/project.py
@@ -7,6 +7,8 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from betty.content import ContentManufacturer
+from betty.enrichers.deriver import Deriver
+from betty.enrichers.wiki import Wiki
 from betty.entity.reference import EntityReference
 from betty.extension import ExtensionManufacturer
 from betty.license import LicenseManufacturer
@@ -24,8 +26,6 @@ from betty.plugins.content.raspberry_mint_incomplete_translation_warning import 
 from betty.plugins.content.raspberry_mint_section import Section, SectionData
 from betty.plugins.content.render import Render, RenderData
 from betty.plugins.content.wikipedia_summary import WikipediaSummary
-from betty.plugins.enricher.deriver import Deriver
-from betty.plugins.enricher.wiki import Wiki
 from betty.plugins.entity.event import Event
 from betty.plugins.entity.person import Person
 from betty.plugins.entity.place import Place
