+++ betty/enrichers/privatizer/__init__.py
@@ -4,12 +4,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Self, final, override
 
+from betty.enrichers.deriver import Deriver
+from betty.enrichers.deriver.jobs import DeriveAncestry
+from betty.enrichers.privatizer.jobs import PrivatizeAncestry
 from betty.factory import Manufacturable
 from betty.load import Enricher, EnricherDefinition
 from betty.locale.localizable.gettext import _
-from betty.plugins.enricher.deriver import Deriver
-from betty.plugins.enricher.deriver.jobs import DeriveAncestry
-from betty.plugins.enricher.privatizer.jobs import PrivatizeAncestry
 from betty.project import Project
 
 if TYPE_CHECKING:
